Switch to 32-bits for internal numbers.
64-bits will be overkill on second-thoughts, esp if we process 60'000+ line source files.
I can always add 64-bit numbers back in later, and even store them only when needed.

+++ src/lib/node.rs
@@ -57,12 +57,12 @@ pub enum NodeKind<'token> {
 #[derive(Debug)]
 pub enum Value {
     /// An integer literal value.
-    Int(i64),
+    Int(i32),
     /// An unsigned literal value as used by binary and hexadecimal
-    /// numbers to represent raw numbers up to 64-bits.
-    UInt(u64),
+    /// numbers to represent raw numbers up to 32-bits.
+    UInt(u32),
     /// A floating point literal value.
-    Float(f64),
+    Float(f32),
 }
 
 /// During building of the `AST`, the methods return either a new `Node` to
